Adding functionalities Add and Remove
