+++ src/index.js
@@ -2,4 +2,5 @@ import { ToDo } from '../modules/ToDoLogic.js';
 
 const toDo = new ToDo();
 
-toDo.loadTasks();
+toDo.executeInputFunctioning();
+toDo.resetTasks();
